test(ingest): put the fifo read-bound case behind a unix build tag
syscall.Mkfifo is not declared on Windows, so a runtime.GOOS skip inside the
shared test file fails to COMPILE there rather than skip. The constraint has to
be a build tag, so the case moves to its own unix-only file.

The call is also watched with a timeout now: a read with no bound on a source
with no end does not return, and that should be reported as the failure it is
instead of stalling the suite until the package timeout.
